pagesの追加（２）
